Fixed camera bug
Issue was with camera inside room prefab.

New issue with falling through floor. This is because the room is assembled as a hollowed cube and doesn't have an actual floor (mesh collider is designed for planes/walls)

Only fixes I have in mind are making transparent planes/walls that overlay the room and have the mesh collider or using box collider (which is buggy and I don't know how it behaves)